requirements-actions.txt
pandas
matplotlib
great_tables
Pillow
instagrapi==2.1.2